+++ cms/src/features/posts/components/PostCard.tsx
@@ -10,6 +10,7 @@ const PostCard: React.FC<Props> = ({ data }) => {
   const { id, title, titleSlug, createdAt } = data;
   const { username } = useParams();
   const postPath = `/${username}/${titleSlug}`;
+  const editPostPath = `/${username}/posts/${data.titleSlug}/edit`;
   // TODO
   // Add
   //  Delete button
@@ -18,10 +19,7 @@ const PostCard: React.FC<Props> = ({ data }) => {
     <li>
       <NavAnchor pathname={postPath} textContent={title} />
       <p>Posted on {createdAt.toString()}</p>
-      <NavAnchor
-        pathname={`/${username}/edit?goto=${postPath}`}
-        textContent="edit"
-      />
+      <NavAnchor pathname={editPostPath} textContent="edit" />
     </li>
   );
 };